Various fixes for auto-generated PropertySet definitions
