+++ Source/Revit.IFC.Common/Utility/IfcSchemaEntityTree.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.IO;
+using System.Xml;
+using System.Xml.Schema;
 using Autodesk.Revit.DB;
 
 namespace Revit.IFC.Common.Utility
@@ -15,6 +17,7 @@ namespace Revit.IFC.Common.Utility
    public class IfcSchemaEntityTree
    {
       static private IDictionary<string, IfcSchemaEntityNode> m_IfcEntityDict = null;
+      static private IDictionary<string, IList<string>> m_PredefTypeEnum = null;
 
       static HashSet<IfcSchemaEntityNode> rootNodes = new HashSet<IfcSchemaEntityNode>();
 
@@ -33,6 +36,31 @@ namespace Revit.IFC.Common.Utility
          }
       }
 
+      static public IDictionary<string, IList<string>> PredefinedTypeEnumDict
+      {
+         get
+         {
+            if (m_PredefTypeEnum == null)
+               m_PredefTypeEnum = new Dictionary<string, IList<string>>();
+            return m_PredefTypeEnum;
+         }
+      }
+
+      static public void AddPredefinedTypeEnum(string enumType, IList<string> enumList)
+      {
+         if (enumType == null || enumList == null || enumList.Count == 0)
+            return;
+
+         if (m_PredefTypeEnum.ContainsKey(enumType))
+         {
+            m_PredefTypeEnum[enumType] = enumList;
+         }
+         else
+         {
+            m_PredefTypeEnum.Add(enumType, enumList);
+         }
+      }
+
       /// <summary>
       /// Reset the static Dictionary and Set. To be done before parsing another IFC schema
       /// </summary>
@@ -44,6 +72,7 @@ namespace Revit.IFC.Common.Utility
 
          // It is a new schema or the first time
          IfcEntityDict.Clear();
+         PredefinedTypeEnumDict.Clear();
          rootNodes.Clear();
          loadedIfcSchemaVersion = "";
       }
@@ -83,10 +112,10 @@ namespace Revit.IFC.Common.Utility
             case IFCVersion.IFC4:
             case IFCVersion.IFC4DTV:
             case IFCVersion.IFC4RV:
-               schemaFile = "IFC4_ADD2.xsd";
+               schemaFile = "IFC4.xsd";
                break;
             default:
-               schemaFile = "IFC4_ADD1.xsd";
+               schemaFile = "IFC4.xsd";
                break;
          }
          return schemaFile;
@@ -105,7 +134,7 @@ namespace Revit.IFC.Common.Utility
             // Process IFCXml schema here, then search for IfcProduct and build TreeView beginning from that node. Allow checks for the tree nodes. Grey out (and Italic) the abstract entity
             string schemaLoc = System.IO.Path.GetDirectoryName(System.Reflection.Assembly.GetCallingAssembly().Location);
             schemaFile = Path.Combine(schemaLoc, schemaFile);
-            FileInfo schemaFileInfo = new FileInfo(schemaFile);
+            FileInfo schemaFileInfo = new FileInfo(schemaFile + ".xsd");
 
             bool newLoad = ProcessIFCXMLSchema.ProcessIFCSchema(schemaFileInfo);
             if (newLoad)
@@ -120,7 +149,7 @@ namespace Revit.IFC.Common.Utility
       /// </summary>
       /// <param name="entityName">the entity name</param>
       /// <param name="parentNodeName">the name of the supertype entity</param>
-      static public void Add(string entityName, string parentNodeName, bool isAbstract = false)
+      static public void Add(string entityName, string parentNodeName, string predefTypeEnum, bool isAbstract = false)
       {
          if (string.IsNullOrEmpty(entityName))
             return;
@@ -152,7 +181,7 @@ namespace Revit.IFC.Common.Utility
          {
             if (parentNode != null)
             {
-               entityNode = new IfcSchemaEntityNode(entityName, parentNode, abstractEntity: isAbstract);
+               entityNode = new IfcSchemaEntityNode(entityName, parentNode, predefTypeEnum, abstractEntity: isAbstract);
                parentNode.AddChildNode(entityNode);
             }
             else
@@ -266,6 +295,46 @@ namespace Revit.IFC.Common.Utility
                {
                   res = entNode;
                   break;
+               }
+            }
+         }
+         return res;
+      }
+
+      /// <summary>
+      /// Collect all the supertype of an entity node
+      /// </summary>
+      /// <param name="entityName">the entity</param>
+      /// <param name="stopNode">array of the stop node(s)</param>
+      /// <returns>List of the supertypes</returns>
+      static public IList<IfcSchemaEntityNode> FindAllSuperTypes(string entityName, params string[] stopNode)
+      {
+         IList<IfcSchemaEntityNode> res = new List<IfcSchemaEntityNode>();
+
+         IfcSchemaEntityNode entNode = Find(entityName);
+
+         if (entNode != null)
+         {
+            // return the list when it reaches the stop node
+            foreach (string stopCond in stopNode)
+               if (entNode.Name.Equals(stopCond, StringComparison.InvariantCultureIgnoreCase))
+                  return res;
+
+            while (true)
+            {
+               entNode = entNode.GetParent();
+               // no more parent node to get
+               if (entNode == null)
+                  break;
+
+               // Stop the search when it reaches the stop node
+               foreach (string stopCond in stopNode)
+                  if (entNode.Name.Equals(stopCond, StringComparison.InvariantCultureIgnoreCase))
+                     break;
+
+               if (entNode != null)
+               {
+                  res.Add(entNode);
                }
             }
          }
