+++ Source/RevitIFCTools/RevitIFCTools.csproj
@@ -58,8 +58,12 @@
     <Reference Include="Antlr4.Runtime, Version=4.6.0.0, Culture=neutral, PublicKeyToken=09abb75b9ed49849, processorArchitecture=MSIL">
       <HintPath>..\..\packages\Antlr4.Runtime.4.6.4\lib\net45\Antlr4.Runtime.dll</HintPath>
     </Reference>
+    <Reference Include="RevitAPI">
+      <HintPath>..\..\..\..\..\Program Files\Autodesk\Revit 2019\RevitAPI.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Data" />
+    <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml" />
     <Reference Include="Microsoft.CSharp" />
@@ -79,11 +83,13 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="IFCEntityAndPsetList.cs" />
     <Compile Include="ParameterExpr\ExprTester.xaml.cs">
       <DependentUpon>ExprTester.xaml</DependentUpon>
     </Compile>
     <Compile Include="ParameterExpr\Logger.cs" />
     <Compile Include="ParameterExpr\ParameterExprParserTest.cs" />
+    <Compile Include="ProcessPsetDefinition.cs" />
     <Compile Include="PropertySet\ComplexProperty.cs" />
     <Compile Include="PropertySet\LanguageType.cs" />
     <Compile Include="PropertySet\NameAlias.cs" />
